requirements.txt
mysqlclient
appdirs==1.4.4
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.52.0
greenlet==1.0.0
httplib2==0.18.1
idna==2.10
importlib-metadata==3.9.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
language-tool-python==2.5.3
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.4
oauthlib==3.1.0
pandas
pipenv==2020.11.15
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn
scipy
selenium
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.3
stripe==2.56.0
threadpoolctl==2.1.0
tqdm==4.55.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
Flask-Migrate
